requirements.txt
# Flask核心框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# 数据库驱动
PyMySQL==1.1.0
cryptography==41.0.4

# Redis和Celery
redis==4.6.0
celery==5.3.1

# AI相关
openai==0.28.1
requests==2.31.0

# 数据处理
pandas==2.0.3
numpy==1.26.4
scikit-learn==1.4.2

# 图表生成
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# 文档生成
jinja2==3.1.2
weasyprint==59.0

# 测试框架
selenium==4.11.2
requests==2.31.0
pytest==7.4.0
pytest-html==3.2.0

# 日志
loguru==0.7.0

# 配置文件
PyYAML==6.0.1

# 工具库
click==8.1.6
werkzeug==2.3.6
markupsafe==2.1.3

# 文件处理
openpyxl==3.1.2
Pillow==10.0.0

# 时间处理
python-dateutil==2.8.2

# 文本处理
jieba==0.42.1

# HTTP客户端
httpx==0.24.1

# 开发工具
flake8==6.0.0
black==23.7.0

# 性能监控
psutil==5.9.5

# 加密
bcrypt==4.0.1

# 邮件
Flask-Mail==0.9.1

# 缓存
Flask-Caching==2.0.2

# API文档
flask-restx==1.1.0

# 环境变量
python-dotenv==1.0.0